pyproject.toml
[tool.poetry]
name = "bq-airflow-dag-generator"
version = "0.1.0"
description = "Generate Airflow DAG from DOT language to execute BigQuery efficiently mainly for AlphaSQL"
authors = ["Matts966 <Matts966@users.noreply.github.com >"]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "^3.6"
networkx = "^2.5"
google-cloud-bigquery = "^2.3.1"
apache-airflow = "^1.10.12"
pydot = "^1.4.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
